Implement Class-based domain event name resolver
It will resolve the event name based on the namespace of the domain event.
This is the full namespace, to avoid conflicts with domain events within
other namespaces but with the same event name.

Note: This is a handy and quick way to resolve the domain event without
the need of extra methods in your event. However, there is a risk involved:

When you would decide to change the namespace of a domain event (in bigger refactor
projects), the resolved name of the domain event will be changed as well. You will
need to implement the NamedDomainEvent in order to keep the old event name.